remove 8-directional movement

+++ core/src/main/input/IPlayerInput.java
@@ -3,7 +3,7 @@ package input;
 public interface IPlayerInput {
 
     boolean isShootPressed();
-    int getX();
-    int getY();
+    float getX();
+    float getY();
 
 }
